pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "castle-bot"
version = "0.1.0"
description = "The Castle — Kalshi paper/demo trading research bot"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.31",
  "cryptography>=41.0",
  "python-dotenv>=1.0",
  "pydantic>=2.6",
  "sqlalchemy>=2.0",
  "tenacity>=8.2",
  "typer>=0.12",
  "rich>=13.7",
  "feedparser>=6.0",
  "numpy>=1.26",
  "pandas>=2.2",
  "PyYAML>=6.0",
]

[project.scripts]
castle = "castle.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
